requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
requests
beautifulsoup4
streamlit
python-dotenv
arxiv
tiktoken